fix rebase

+++ contracts/hts-precompile/HederaTokenService.sol
@@ -689,7 +689,7 @@ abstract contract HederaTokenService {
     /// @param account The account address to revoke kyc
     /// @param amount The number of tokens to wipe
     /// @return responseCode The response code for the status of the request. SUCCESS is 22.
-    function delegateWipeTokenAccount(address token, address account, uint32 amount) internal returns (int responseCode)
+    function delegateWipeTokenAccount(address token, address account, int64 amount) internal returns (int responseCode)
     {
         (bool success, bytes memory result) = precompileAddress.delegatecall(
             abi.encodeWithSelector(IHederaTokenService.wipeTokenAccount.selector, token, account, amount));
